flask_py38_requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
absl-py==2.1.0
alabaster
alembic==1.13.1
amqp==5.2.0
anyio==4.4.0
APScheduler==3.11.0
arabic-reshaper==3.0.0
arrow
asn1crypto==1.5.1
astroid
asttokens
astunparse==1.6.3
async-timeout==4.0.3
atomicwrites==1.4.0
attrs
autopep8
Babel
backcall
backports.zoneinfo==0.2.1
bcrypt
beautifulsoup4==4.11.1
binaryornot
black
bleach
blinker==1.8.2
blis==0.7.11
boto3==1.37.38
botocore==1.37.38
Brotli
cachetools==5.5.0
catalogue==2.0.10
catboost==1.2.8
celery==5.4.0
certifi
cffi
chardet
charset-normalizer
ci-info==0.3.0
click
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle
colorama
colorlog==6.8.2
comm
confection==0.1.5
configobj==5.0.8
configparser==7.1.0
contourpy==1.1.1
cookiecutter
cryptography==45.0.5
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
debugpy
decorator
defusedxml
diff-match-patch
dill
distro==1.9.0
docstring-to-markdown
docutils
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing
Faker==33.1.0
fastjsonschema
filelock==3.16.0
flake8
Flask==2.3.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fpdf2==2.8.1
frozenlist==1.4.1
fsspec==2025.3.0
future==1.0.0
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.68.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperopt==0.2.7
idna
imagesize
imbalanced-learn==0.12.4
importlib-metadata
importlib-resources
inflection==0.5.1
intervaltree
ipykernel
ipython
isodate==0.6.1
isort
itsdangerous==2.2.0
jaraco.classes
jedi
jellyfish
Jinja2
jiter==0.5.0
jmespath==1.0.1
joblib==1.1.1
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab-pygments
keyring
kiwisolver==1.4.7
kombu==5.4.2
langcodes==3.4.0
language_data==1.2.0
lazy-object-proxy
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.41.1
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py
MarkupSafe
matplotlib==3.7.5
matplotlib-inline
mccabe
mdurl
mistune
more-itertools
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions
narwhals==1.42.1
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numba==0.58.1
numpy==1.24.4
numpydoc
oauthlib==3.2.2
openai==0.28.0
openai-whisper
openpyxl==3.1.2
opt_einsum==3.4.0
optuna==3.6.1
oscrypto==1.3.0
packaging
pandas==2.0.3
pandocfilters
paramiko
parso
pathlib==1.0.1
pathlib_abc==0.1.1
pathspec
pathy==0.11.0
pdf2image==1.17.0
pdfminer.six==20240706
pexpect
pickleshare
pillow==10.4.0
pkgutil_resolve_name
platformdirs
plotly==6.2.0
pluggy
ply==3.11
preshed==3.0.9
prompt-toolkit
protobuf==4.25.5
prov==2.0.1
psutil
psycopg2-binary==2.9.3
ptyprocess
pure-eval
py4j==0.10.9.7
pyaml==24.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle
pycparser
pydantic==1.10.17
pydocstyle
pydot==3.0.1
pyflakes
Pygments
pyHanko==0.27.1
pyhanko-certvalidator==0.26.8
pylint
pylint-venv
pyls-spyder==0.4.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyNaCl
pyparsing==3.1.4
pypdf==5.9.0
PyPDF2==3.0.1
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-sip
PyQtWebEngine==5.15.6
PySocks
pytesseract==0.3.10
python-bidi==0.6.6
python-dateutil
python-decouple==3.5
python-docx==1.1.2
python-dotenv==1.0.0
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
python-slugify
pytoolconfig
pytz==2023.3
pywin32==305
pywin32-ctypes
pyxnat==1.6.2
PyYAML
pyzmq
QDarkStyle
qrcode==7.4.2
qstylizer
QtAwesome
qtconsole
QtPy
rdflib==6.3.2
redis==5.0.8
referencing
regex==2024.11.6
reportlab==4.2.5
requests
requests-oauthlib==2.0.0
rich
rope
rpds-py
rsa==4.9
Rtree
s3transfer==0.11.5
scikit-learn==1.3.2
scikit-optimize==0.9.0
scipy==1.10.1
seaborn==0.13.2
shap==0.44.1
simplejson==3.19.3
sip
six
slicer==0.0.7
smart-open==6.4.0
sniffio==1.3.1
snowballstemmer
sortedcontainers
soupsieve
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy==2.0.36
srsly==2.4.8
stack-data
svglib==1.5.1
sympy==1.13.3
termcolor==2.4.0
text-unidecode
textdistance
thinc==8.1.12
threadpoolctl==3.5.0
three-merge
tiktoken==0.7.0
tinycss2
tomli
tomlkit
torch==2.4.1
tornado
tqdm==4.66.4
traitlets
traits==6.3.2
typer==0.7.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson
Unidecode
uritools==4.0.3
urllib3==1.26.20
vine==5.1.0
wasabi==1.1.3
watchdog==4.0.1
wcwidth
webencodings==0.5.1
Werkzeug==2.3.7
whatthepatch
win-inet-pton
wrapt
xgboost==2.1.3
xhtml2pdf==0.2.17
XlsxWriter==1.4.4
yapf
yarl==1.9.4
zipp


